add DELETE :bucket/:entry/q endpoint
