+++ integration_tests/data_check/uploader.py
@@ -7,8 +7,8 @@ from reduct import Client, Bucket, ServerInfo
 from os import getenv
 import random
 
-NUMBER_OF_ENTRIES = int(getenv("NUMBER_OF_ENTRIES", 16))
-NUMBER_OF_RECORDS = int(getenv("NUMBER_OF_RECORDS", 512))
+NUMBER_OF_ENTRIES = int(getenv("NUMBER_OF_ENTRIES", 4))
+NUMBER_OF_RECORDS = int(getenv("NUMBER_OF_RECORDS", 2048))
 MAX_BLOB_SIZE = 1024 * 1024 * int(getenv("MAX_BLOB_SIZE", 1))
 
 BLOB = random.randbytes(MAX_BLOB_SIZE)
